Implement run-length encoding compression scheme
Graphics are now compressed with a simple form of run-length encoding!
This means that our "trident" executable can now shrink from ~400KiB
down to ~35KiB!!  The savings are massive for the simple USB trident
pixmap!  Graphics stored in RLE format are now represented in chunks
consisting of two `uint32_t` values: RGBA pixel value and number of
repetitions.

Also in this commit are some changes to the configuration file JSON
format.  All pixmaps are to be stored in the "gfx" array as dictionaries
consisting of a name, dimensions [W,H], the source pixmap filename, and
an enumerated type (see trident.h:6).