configure: remove need for legacy trident.h symlink in EFI
The future of the machine image proper for the targetdisk project lives
on in td-efi project.  The Linux framebuffer API graphics demo in this
repository largely lives on for posterity, in the hopes that someone
somewhere someday might find it educational.  The configure script and
its associated paraphernalia are still used by external projects.